refactor: extract authData optimistic lock to shared util
